html/editing/editing-0/autocorrection/autocorrection.html

<!DOCTYPE html>
<html>
<link rel="help" href="https://html.spec.whatwg.org/multipage/interaction.html#autocorrection">
<body>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

test(function() {
    assert_true('autocorrect' in document.createElement('input'));
}, "Test that the autocorrect is available on HTMLInputElement.")

test(function() {
    assert_true('autocorrect' in document.createElement('textarea'));
}, "Test that the autocorrect is available on HTMLTextAreaElement.")

test(function() {
    assert_true('autocorrect' in document.createElement('div'));
}, "Test that the autocorrect is available on div.")

test(function() {
    assert_true('autocorrect' in document.createElement('form'));
}, "Test that the autocorrect is available on form.")

test(function() {
  var elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('form') ];

  elements.forEach(function(e) {
    e.autocorrect = true;
    assert_true(e.autocorrect);
    assert_equals(e.getAttribute('autocorrect'), 'on');

    e.autocorrect = 'hello';
    assert_true(e.autocorrect);
    assert_equals(e.getAttribute('autocorrect'), 'on');

    e.autocorrect = false;
    assert_false(e.autocorrect);
    assert_equals(e.getAttribute('autocorrect'), 'off');

    e.autocorrect = 0;
    assert_false(e.autocorrect);
    assert_equals(e.getAttribute('autocorrect'), 'off');
  });
}, "Test setting the autocorrect boolean attribute.");

test(function() {
  var elements = [ document.createElement('input'),
                   document.createElement('textarea'),
                   document.createElement('div'),
                   document.createElement('form') ];

  elements.forEach(function(e) {
    e.setAttribute('autocorrect', 'on');
    assert_equals(e.getAttribute('autocorrect'), 'on');
    assert_true(e.autocorrect);

    e.setAttribute('autocorrect', 'ON');
    assert_equals(e.getAttribute('autocorrect'), 'ON');
    assert_true(e.autocorrect);

    e.setAttribute('autocorrect', 'off');
    assert_equals(e.getAttribute('autocorrect'), 'off');
    assert_false(e.autocorrect);

    e.setAttribute('autocorrect', 'OFF');
    assert_equals(e.getAttribute('autocorrect'), 'OFF');
    assert_false(e.autocorrect);

    e.setAttribute('autocorrect', 'invalid_value');
    assert_equals(e.getAttribute('autocorrect'), 'invalid_value');
    assert_true(e.autocorrect);

    e.setAttribute('autocorrect', '');
    assert_equals(e.getAttribute('autocorrect'), '');
    assert_true(e.autocorrect);

    e.removeAttribute('autocorrect');
    assert_false(e.hasAttribute('autocorrect'));
    assert_true(e.autocorrect);
  });
}, "Test setting the autocorrect attribute using setAttribute.")

test(function() {
  const testData = [
    {
      formValue: null,
      formElementValue: null,
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: null,
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: null,
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: null,
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: 'off',
      formElementValue: null,
      inheritedResult: false,
      uninheritedResult: true,
    },
    {
      formValue: 'off',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: 'on',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: 'off',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'off',
      formElementValue: 'foo',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'off',
      formElementValue: 'bar',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'off',
      formElementValue: '',
      inheritedResult: false,
      uninheritedResult: true,
    },
    {
      formValue: 'off',
      formElementValue: '',
      inheritedResult: false,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: 'foo',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: 'bar',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: '',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'on',
      formElementValue: '',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'foo',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: 'bar',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: '',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: '',
      formElementValue: 'off',
      inheritedResult: false,
      uninheritedResult: false,
    },
    {
      formValue: 'foo',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: 'bar',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: '',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    },
    {
      formValue: '',
      formElementValue: 'on',
      inheritedResult: true,
      uninheritedResult: true,
    }
  ];

  const formElements = [
    {element: 'button', inherits: true},
    {element: 'fieldset', inherits: true},
    {element: 'img', inherits: false},
    {element: 'input', inherits: true},
    {element: 'object', inherits: false},
    {element: 'output', inherits: true},
    {element: 'select', inherits: true},
    {element: 'textarea', inherits: true},
  ];

  const form = document.createElement('form');
  form.id = 'form';
  document.body.appendChild(form);

  testData.forEach(data => {
    form.removeAttribute('autocorrect');

    if (data.formValue !== null) {
      form.setAttribute('autocorrect', data.formValue);
    }

    formElements.forEach(elementData => {
      const element = document.createElement(elementData.element);
      form.appendChild(element);

      const element2 = document.createElement(elementData.element);
      element2.setAttribute('form', 'form');
      document.body.appendChild(element2);

      if (data.formElementValue !== null) {
        element.setAttribute('autocorrect', data.formElementValue);
        element2.setAttribute('autocorrect', data.formElementValue);
      }

      const descriptionSuffix = 'with "' + data.formValue
          + '" and form element with "'+ data.formElementValue + '"';

      if (elementData.inherits) {
        assert_equals(element.autocorrect, data.inheritedResult,
                      `${elementData.element} element with form parent `
                          + `${descriptionSuffix}`);
        assert_equals(element2.autocorrect, data.inheritedResult,
                      `${elementData.element} element with form owner attribute`
                          + ` set ${descriptionSuffix}`);
      } else {
        assert_equals(element.autocorrect, data.uninheritedResult,
                      `${elementData.element} element with form parent `
                          + `${descriptionSuffix}`);
        assert_equals(element2.autocorrect, data.uninheritedResult,
                      `${elementData.element} element with form owner attribute`
                          + `set ${descriptionSuffix}`);
      }
    });
  });
}, "Test inheriting autocorrection from a form.")

test(function() {
  const editingHost = document.createElement("div");
  const container = document.createElement("br");
  editingHost.contentEditable = true;
  editingHost.appendChild(container);
  document.body.appendChild(editingHost);

  editingHost.autocorrect = false;
  container.autocorrect = true;

  assert_false(editingHost.autocorrect);
  assert_true(container.autocorrect);

  editingHost.autocorrect = true;
  container.autocorrect = false;

  assert_true(editingHost.autocorrect);
  assert_false(container.autocorrect);
}, "Test autocorrection in an editing host.")

</script>
</body>
</html>
